+++ Editor/Drawers/LeftToggleAttributeDrawer.cs
@@ -1,0 +1,29 @@
+﻿using SaintsField.Editor.Core;
+using UnityEditor;
+using UnityEngine;
+
+namespace SaintsField.Editor.Drawers
+{
+    [CustomPropertyDrawer(typeof(LeftToggleAttribute))]
+    public class LeftToggleAttributeDrawer: SaintsPropertyDrawer
+    {
+        protected override float GetFieldHeight(SerializedProperty property, GUIContent label,
+            ISaintsAttribute saintsAttribute, bool hasLabelWidth)
+        {
+            return EditorGUIUtility.singleLineHeight;
+        }
+
+        protected override void DrawField(Rect position, SerializedProperty property, GUIContent label, ISaintsAttribute saintsAttribute)
+        {
+            // ReSharper disable once ConvertToUsingDeclaration
+            using(EditorGUI.ChangeCheckScope changed = new EditorGUI.ChangeCheckScope())
+            {
+                bool result = EditorGUI.ToggleLeft(position, label, property.boolValue);
+                if (changed.changed)
+                {
+                    property.boolValue = result;
+                }
+            }
+        }
+    }
+}