+++ Runtime/LeftToggleAttribute.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace SaintsField
+{
+    public class LeftToggleAttribute: PropertyAttribute, ISaintsAttribute
+    {
+        public SaintsAttributeType AttributeType => SaintsAttributeType.Field;
+        public string GroupBy => "__LABEL_FIELD__";
+    }
+}