+++ Samples~/Scripts/LeftToggleExample.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+namespace SaintsField.Samples.Scripts
+{
+    public class LeftToggleExample: MonoBehaviour
+    {
+        [LeftToggle] public bool myToggle;
+        // To use with `RichLabel`, you need to add 6 spaces ahead as a hack
+        [LeftToggle, RichLabel("      <color=green><label />")] public bool richToggle;
+    }
+}