+++ Samples~/Scripts/SceneExample.cs
@@ -4,7 +4,7 @@ namespace SaintsField.Samples.Scripts
 {
     public class SceneExample: MonoBehaviour
     {
-        [Scene] public int _sceneInt;
-        [Scene] public string _sceneString;
+        [Scene] public int sceneInt;
+        [Scene, BelowRichLabel(nameof(sceneStr), true)] public string sceneStr;
     }
 }